+++ src/utils/date/index.ts
@@ -1,0 +1,2 @@
+export * from './dayjsJa'
+export * from './scripts'
